New Game
I am trying to make a new game with different enemies instead of characters.

For the New Game file, I found a lot of the code off the internet as it didn't really work when I copied it from the pre-existing game. However, this doesn't really work either, as it just makes my screen really glitchy and I have to close VS Code. Do you know why it is going wrong?